backend/requirements.txt
Django==5.1
djangorestframework==3.15.2
djoser==2.2.3
drf-yasg==1.21.7
python-dotenv==1.0.1
setuptools==72.2.0
